code/requirements.txt
Django
psycopg2
django-rest-framework
drf-yasg